Promote ExtractorMessage and ExtractorMessageId to be public API and add docs
