+++ apps/api-extractor/src/index.ts
@@ -12,6 +12,8 @@
 export { ReleaseTag } from './aedoc/ReleaseTag';
 
 export { Extractor, IAnalyzeProjectOptions, IExtractorOptions } from './api/Extractor';
+export { ExtractorMessage, ExtractorMessageCategory } from './api/ExtractorMessage';
+export { ExtractorMessageId } from './api/ExtractorMessageId';
 export {
   IExtractorTsconfigCompilerConfig,
   IExtractorRuntimeCompilerConfig,
